config/pyproject.toml
[project]
name = "config"
version = "0.1.0"
description = "This module provides functionality for managing configuration settings in Python applications."
authors = ["Kyle Rose <kylemail1946@gmail.com>"]
maintainers = ["Kyle Rose <kylemail1946@gmail.com>"]
repository = "https://github.com/the-ride-never-ends/config"
readme = "README.md"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["config"]